requirements.txt
django
djangorestframework
django-sendfile
requests
lxml